emailer/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
awscli==1.33.26
beautifulsoup4==4.12.2
boto3==1.34.144
botocore==1.34.144
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
docutils==0.16
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.0
pandas==2.0.2
pyasn1==0.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
